requirements.txt
fpdf
reportlab 
Pillow 
pytesseract
pdf2image
pymupdf
